This file is the diagram in SAS enterprise miner
In this data mining report, we will be going to consider the Framingham dataset for our analysis to find out the factors contributing to the Cardio Vascular Disease (CVD). Using SAS Enterprise Miner we will conduct this analysis and try to compare different methods to mine the data. Also, the comparison between different Neural networks, Regression, Decision trees, and other research nodes, will be the key aspects of this report. The dataset consists of different parameters which might or not contribute to the disease and we will explicitly dig down to those factors. 
Any disorder that affects the heart or blood vessels is referred to as cardiovascular disease (CVD). It is usually linked to the accumulation of fatty deposits in the arteries and an increased risk of blood clots. It's also been connected to damage to arteries in organs including the brain, heart, kidneys, and eyes. These are the reasons, why this exercise is important for society and mankind. This dataset has a lot of potentials and can be used for further research also.